Added 8bit ALU operations, still need to implement (HL) variations

+++ include/forwarddecls.h
@@ -1,5 +1,7 @@
+#include "spdlog/spdlog.h"
+
 class CPU;
 
 typedef unsigned char Ticks;
 typedef unsigned char Byte;
-typedef unsigned int Word;
+typedef unsigned short Word;
